Add files via upload
While watching the old spy movie, we were wondering how the spy used to send the secret messages. Therefore, we looked into the historical documents which mentioned several encrypting methods used since ancient Rome. Then we hoped to use modern computing technology to simplify the encoding and decoding processes. Thus, we built a python program to encrypt and decrypt message based on shift encoding and table encoding. With the combination of those two methods, the encrypted messages will be almost impossible for decrypting without the corresponding keys and programs. We developed such program to help the beginner for computer science to understand the mechanism behind the encoding system and to create a simply and interesting way for level entry users to communicate with each other using secret messages without going through troubles with encoding and decoding. The last but not the least, we did it to honor the old spy movies.